# [Swift] Implements union strings