This PR includes the support for string within unions in swift. 
This was done by replacing `Offset<UOffset>, and Offset<String>` to only be `Offset`

Master:
```
done 500_000: 0.0805s
done 10 str:  0.0006s
done 100 str: 0.0016s
done 3M strc: 0.7577s
```

This implementation:
```
done 500_000: 0.0745s
done 10 str:  0.0005s
done 100 str: 0.0015s
done 3M strc: 0.6674s
```
As you can see here the benchmarks for this branch sped up